Implement restaurant authorization and ownership features
- Added `IRestaurantAuthorizationService` for managing authorization logic.
- Updated command handlers to include authorization checks.
- Modified `CreateRestaurantCommandHandler` to set the restaurant owner.
- Enhanced `Restaurant` and `User` entities with ownership properties.
- Created migration to update the database schema for ownership.
- Removed unused `Validators` folder references from the project.
- Introduced `ResourceOperation` enum and `ForbiddenException` class.
- Updated `DbContext` and model snapshot to reflect new relationships.

+++ Restaurants.Applications/Dishes/Commands/DeleteAllDishForRestaurant/DeleteAllDishForRestaurantCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Applications.Dishes.Commands.DeleteAllDishForRestaurant
@@ -8,6 +9,7 @@ namespace Restaurants.Applications.Dishes.Commands.DeleteAllDishForRestaurant
     public class DeleteAllDishForRestaurantCommandHandler(
             ILogger<DeleteAllDishForRestaurantCommandHandler> logger,
             IDishesRepository dishesRepository,
+            IRestaurantAuthorizationService restaurantAuthorizationService,
             IRestaurantsRepository restaurantsRepository
         ) 
         
@@ -24,6 +26,11 @@ namespace Restaurants.Applications.Dishes.Commands.DeleteAllDishForRestaurant
 
             if (restaurant.Dishes == null || !restaurant.Dishes.Any())
                 throw new NotFoundException(nameof(restaurant.Dishes), request.RestaurantId.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Update))
+            {
+                throw new ForbiddenException();
+            }
 
             await dishesRepository.DeleteAllDishAsync(restaurant.Dishes);
